refactor(commands): move result serialization to the command surface
The `serialize*` helpers render command results for the CLI and MCP surfaces,
not for the daemon; only the two deployment target resolvers were shared with
daemon handlers. The serializers move to `src/commands/output/` beside the
formatters that call them, and the resolvers move to `src/core` below both
owners, so `src/commands` and `src/mcp` no longer reach into `src/daemon` for
them. The MCP edge becomes a static import; the `src/commands` edges stay
function-scoped because `src/cli.ts` eagerly evaluates both output modules.
