+++ src/daemon/handlers/session-app-deployment.ts
@@ -12,7 +12,7 @@ import type { BindDeviceRuntime, InspectDeviceRuntimeFacts } from '../request-ru
 import { SessionStore } from '../session-store.ts';
 import type { DaemonRequest, DaemonResponse, SessionState } from '../types.ts';
 import { resolvePayloadInput } from '../../core/payload-input.ts';
-import { resolveDeployResultTarget } from '../result-serialization.ts';
+import { resolveDeployResultTarget } from '../../core/deploy-result-target.ts';
 import { withSuccessText } from '@agent-device/kernel/success-text';
 import { recordSessionAction } from '../session-action-recorder.ts';
 import { errorResponse } from '../response.ts';
