+++ src/daemon/handlers/session-app-source-deployment.ts
@@ -16,7 +16,7 @@ import { resolveInstallSource } from '../install-source-resolution.ts';
 import type { BindDeviceRuntime, InspectDeviceRuntimeFacts } from '../request-runtime-binding.ts';
 import { SessionStore } from '../session-store.ts';
 import type { DaemonRequest, DaemonResponse, SessionState } from '../types.ts';
-import { resolveInstallFromSourceResultTarget } from '../result-serialization.ts';
+import { resolveInstallFromSourceResultTarget } from '../../core/deploy-result-target.ts';
 import { withSuccessText } from '@agent-device/kernel/success-text';
 import { recordSessionAction } from '../session-action-recorder.ts';
 import { resolveCommandDevice } from '../session-device-resolution.ts';
